Make it looks nice and add responsive
